Add files via upload
this codes are used to find a volume and percentage of a water present in a tank. Here I used truncated cone as the shape of the tank. 
The python file contains the code for interfacing sql with arduino code and viewing it in the xampp. The sketch_may15a.ino contains the code for esp32 module to connect and operate an ultrasonic sensor.